\name{adm}
\docType{data}
\alias{adm}
\title{Graduate School Admissions Data of University Students}
\description{
  This data set gives the graduate school admission status and university academic information of 400 students.
  \dQuote{major}
  UCLA Institute for Digital Research and Education example data.
}
\usage{adm}
\format{A dataframe containing 400 observations.}
\source{read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")}
\keyword{datasets}
